requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
groq>=0.3.0
streamlit>=1.32.0
python-dotenv==0.21.1
requests==2.28.2
pandas>=1.5.0
pydantic>=1.8.0